+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#27](https://github.com/LaxarJS/laxar-angular-adapter/issues/27): initialize `$rootScope.i18n` exactly once
 - [#34](https://github.com/LaxarJS/laxar-angular-adapter/issues/34): project: updated dev-dependencies, upgraded to webpack 2
 - [#33](https://github.com/LaxarJS/laxar-angular-adapter/issues/33): fixed imports of `laxar-widget-service-mocks`
 - [#32](https://github.com/LaxarJS/laxar-angular-adapter/issues/32): adapted to laxar API change (LaxarJS/laxar#413)