+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.0.0-alpha.0
+
 - [#25](https://github.com/LaxarJS/laxar-angular-adapter/issues/25): fixed testability by running $rootScope.$digest on heartbeat
     + **BREAKING CHANGE:** see ticket for details
 - [#27](https://github.com/LaxarJS/laxar-angular-adapter/issues/27): initialize `$rootScope.i18n` exactly once